+++ api/grpcserver/v1/activation_service_test.go
@@ -1,4 +1,4 @@
-package grpcserver
+package v1
 
 import (
 	"context"
