+++ api/grpcserver/v1/events.go
@@ -1,4 +1,4 @@
-package grpcserver
+package v1
 
 import (
 	"context"
