+++ api/grpcserver/v1/interface.go
@@ -1,4 +1,4 @@
-package grpcserver
+package v1
 
 import (
 	"context"
@@ -15,7 +15,7 @@ import (
 	"github.com/spacemeshos/go-spacemesh/system"
 )
 
-//go:generate mockgen -typed -package=grpcserver -destination=./mocks.go -source=./interface.go
+//go:generate mockgen -typed -package=v1 -destination=./mocks.go -source=./interface.go
 
 // networkInfo interface.
 type networkInfo interface {
