+++ api/grpcserver/v1/mocks.go
@@ -3,11 +3,11 @@
 //
 // Generated by this command:
 //
-//	mockgen -typed -package=grpcserver -destination=./mocks.go -source=./interface.go
+//	mockgen -typed -package=v1 -destination=./mocks.go -source=./interface.go
 //
 
-// Package grpcserver is a generated GoMock package.
-package grpcserver
+// Package v1 is a generated GoMock package.
+package v1
 
 import (
 	context "context"
