+++ api/grpcserver/v1/node_service.go
@@ -1,4 +1,4 @@
-package grpcserver
+package v1
 
 import (
 	"context"
