+++ api/grpcserver/v1/post_service.go
@@ -1,4 +1,4 @@
-package grpcserver
+package v1
 
 import (
 	"context"
