+++ api/grpcserver/v2alpha1/node.go
@@ -11,10 +11,6 @@ import (
 	"github.com/spacemeshos/go-spacemesh/timesync"
 )
 
-const (
-	Node = "node_v2alpha1"
-)
-
 // nodePeerCounter is an api to get current peer count.
 type nodePeerCounter interface {
 	PeerCount() uint64
