+++ node/node.go
@@ -38,6 +38,7 @@ import (
 
 	"github.com/spacemeshos/go-spacemesh/activation"
 	"github.com/spacemeshos/go-spacemesh/api/grpcserver"
+	v1 "github.com/spacemeshos/go-spacemesh/api/grpcserver/v1"
 	"github.com/spacemeshos/go-spacemesh/api/grpcserver/v2alpha1"
 	"github.com/spacemeshos/go-spacemesh/api/grpcserver/v2beta1"
 	"github.com/spacemeshos/go-spacemesh/atxsdata"
@@ -1103,7 +1104,7 @@ func (app *App) initServices(ctx context.Context) error {
 
 	nipostBuilder, err := activation.NewNIPostBuilder(
 		app.localDB,
-		grpcPostService.(*grpcserver.PostService),
+		grpcPostService.(*v1.PostService),
 		nipostLogger,
 		app.Config.POET,
 		app.clock,
@@ -1498,15 +1499,15 @@ func (app *App) grpcService(svc grpcserver.Service, lg log.Log) (grpcserver.Serv
 
 	switch svc {
 	case grpcserver.Debug:
-		service := grpcserver.NewDebugService(app.db, app.conState, app.host, app.hOracle, app.loggers)
+		service := v1.NewDebugService(app.db, app.conState, app.host, app.hOracle, app.loggers)
 		app.grpcServices[svc] = service
 		return service, nil
 	case grpcserver.GlobalState:
-		service := grpcserver.NewGlobalStateService(app.mesh, app.conState)
+		service := v1.NewGlobalStateService(app.mesh, app.conState)
 		app.grpcServices[svc] = service
 		return service, nil
 	case grpcserver.Mesh:
-		service := grpcserver.NewMeshService(
+		service := v1.NewMeshService(
 			app.cachedDB,
 			app.mesh,
 			app.conState,
@@ -1520,7 +1521,7 @@ func (app *App) grpcService(svc grpcserver.Service, lg log.Log) (grpcserver.Serv
 		app.grpcServices[svc] = service
 		return service, nil
 	case grpcserver.Node:
-		service := grpcserver.NewNodeService(
+		service := v1.NewNodeService(
 			app.host,
 			app.mesh,
 			app.clock,
@@ -1531,7 +1532,7 @@ func (app *App) grpcService(svc grpcserver.Service, lg log.Log) (grpcserver.Serv
 		app.grpcServices[svc] = service
 		return service, nil
 	case grpcserver.Admin:
-		service := grpcserver.NewAdminService(app.db, app.Config.DataDir(), app.host)
+		service := v1.NewAdminService(app.db, app.Config.DataDir(), app.host)
 		app.grpcServices[svc] = service
 		return service, nil
 	case grpcserver.Smesher:
@@ -1544,10 +1545,10 @@ func (app *App) grpcService(svc grpcserver.Service, lg log.Log) (grpcserver.Serv
 		if err != nil {
 			return nil, err
 		}
-		service := grpcserver.NewSmesherService(
+		service := v1.NewSmesherService(
 			app.atxBuilder,
 			app.postSupervisor,
-			postService.(*grpcserver.PostService),
+			postService.(*v1.PostService),
 			app.Config.API.SmesherStreamInterval,
 			app.Config.SMESHING.Opts,
 			sig,
@@ -1555,7 +1556,7 @@ func (app *App) grpcService(svc grpcserver.Service, lg log.Log) (grpcserver.Serv
 		app.grpcServices[svc] = service
 		return service, nil
 	case grpcserver.Post:
-		service := grpcserver.NewPostService(app.addLogger(PostServiceLogger, lg).Zap())
+		service := v1.NewPostService(app.addLogger(PostServiceLogger, lg).Zap())
 		isCoinbaseSet := app.Config.SMESHING.CoinbaseAccount != ""
 		if !isCoinbaseSet {
 			lg.Warning("coinbase account is not set, connections from remote post services will be rejected")
@@ -1564,11 +1565,11 @@ func (app *App) grpcService(svc grpcserver.Service, lg log.Log) (grpcserver.Serv
 		app.grpcServices[svc] = service
 		return service, nil
 	case grpcserver.PostInfo:
-		service := grpcserver.NewPostInfoService(app.atxBuilder)
+		service := v1.NewPostInfoService(app.atxBuilder)
 		app.grpcServices[svc] = service
 		return service, nil
 	case grpcserver.Transaction:
-		service := grpcserver.NewTransactionService(
+		service := v1.NewTransactionService(
 			app.db,
 			app.host,
 			app.mesh,
@@ -1579,117 +1580,117 @@ func (app *App) grpcService(svc grpcserver.Service, lg log.Log) (grpcserver.Serv
 		app.grpcServices[svc] = service
 		return service, nil
 	case grpcserver.Activation:
-		service := grpcserver.NewActivationService(app.cachedDB, types.ATXID(app.Config.Genesis.GoldenATX()))
+		service := v1.NewActivationService(app.cachedDB, types.ATXID(app.Config.Genesis.GoldenATX()))
 		app.grpcServices[svc] = service
 		return service, nil
-	case v2alpha1.Activation:
+	case grpcserver.ActivationV2Alpha1:
 		service := v2alpha1.NewActivationService(app.apiDB, types.ATXID(app.Config.Genesis.GoldenATX()))
 		app.grpcServices[svc] = service
 		return service, nil
-	case v2alpha1.ActivationStream:
+	case grpcserver.ActivationStreamV2Alpha1:
 		service := v2alpha1.NewActivationStreamService(app.apiDB)
 		app.grpcServices[svc] = service
 		return service, nil
-	case v2alpha1.Reward:
+	case grpcserver.RewardV2Alpha1:
 		service := v2alpha1.NewRewardService(app.apiDB)
 		app.grpcServices[svc] = service
 		return service, nil
-	case v2alpha1.RewardStream:
+	case grpcserver.RewardStreamV2Alpha1:
 		service := v2alpha1.NewRewardStreamService(app.apiDB)
 		app.grpcServices[svc] = service
 		return service, nil
-	case v2alpha1.Malfeasance:
+	case grpcserver.MalfeasanceV2Alpha1:
 		service := v2alpha1.NewMalfeasanceService(app.apiDB, app.malfeasance2Handler, app.malfeasanceHandler)
 		app.grpcServices[svc] = service
 		return service, nil
-	case v2alpha1.MalfeasanceStream:
+	case grpcserver.MalfeasanceStreamV2Alpha1:
 		service := v2alpha1.NewMalfeasanceStreamService(app.apiDB, app.malfeasance2Handler, app.malfeasanceHandler)
 		app.grpcServices[svc] = service
 		return service, nil
-	case v2alpha1.Network:
+	case grpcserver.NetworkV2Alpha1:
 		service := v2alpha1.NewNetworkService(
 			app.clock.GenesisTime(),
 			app.Config,
 		)
 		app.grpcServices[svc] = service
 		return service, nil
-	case v2alpha1.Node:
+	case grpcserver.NodeV2Alpha1:
 		service := v2alpha1.NewNodeService(app.host, app.mesh, app.clock, app.syncer)
 		app.grpcServices[svc] = service
 		return service, nil
-	case v2alpha1.Layer:
+	case grpcserver.LayerV2Alpha1:
 		service := v2alpha1.NewLayerService(app.apiDB)
 		app.grpcServices[svc] = service
 		return service, nil
-	case v2alpha1.LayerStream:
+	case grpcserver.LayerStreamV2Alpha1:
 		service := v2alpha1.NewLayerStreamService(app.apiDB)
 		app.grpcServices[svc] = service
 		return service, nil
-	case v2alpha1.Transaction:
+	case grpcserver.TransactionV2Alpha1:
 		service := v2alpha1.NewTransactionService(app.apiDB, app.conState, app.syncer, app.txHandler, app.host)
 		app.grpcServices[svc] = service
 		return service, nil
-	case v2alpha1.TransactionStream:
+	case grpcserver.TransactionStreamV2Alpha1:
 		service := v2alpha1.NewTransactionStreamService()
 		app.grpcServices[svc] = service
 		return service, nil
-	case v2alpha1.Account:
+	case grpcserver.AccountV2Alpha1:
 		service := v2alpha1.NewAccountService(app.apiDB, app.conState)
 		app.grpcServices[svc] = service
 		return service, nil
 	// v2beta1
-	case v2beta1.Activation:
+	case grpcserver.ActivationV2Beta1:
 		service := v2beta1.NewActivationService(app.apiDB, types.ATXID(app.Config.Genesis.GoldenATX()))
 		app.grpcServices[svc] = service
 		return service, nil
-	case v2beta1.ActivationStream:
+	case grpcserver.ActivationStreamV2Beta1:
 		service := v2beta1.NewActivationStreamService(app.apiDB)
 		app.grpcServices[svc] = service
 		return service, nil
-	case v2beta1.Reward:
+	case grpcserver.RewardV2Beta1:
 		service := v2beta1.NewRewardService(app.apiDB)
 		app.grpcServices[svc] = service
 		return service, nil
-	case v2beta1.RewardStream:
+	case grpcserver.RewardStreamV2Beta1:
 		service := v2beta1.NewRewardStreamService(app.apiDB)
 		app.grpcServices[svc] = service
 		return service, nil
-	case v2beta1.Malfeasance:
+	case grpcserver.MalfeasanceV2Beta1:
 		service := v2beta1.NewMalfeasanceService(app.apiDB, app.malfeasance2Handler, app.malfeasanceHandler)
 		app.grpcServices[svc] = service
 		return service, nil
-	case v2beta1.MalfeasanceStream:
+	case grpcserver.MalfeasanceStreamV2Beta1:
 		service := v2beta1.NewMalfeasanceStreamService(app.apiDB, app.malfeasance2Handler, app.malfeasanceHandler)
 		app.grpcServices[svc] = service
 		return service, nil
-	case v2beta1.Network:
+	case grpcserver.NetworkV2Beta1:
 		service := v2beta1.NewNetworkService(
 			app.clock.GenesisTime(),
 			app.Config,
 		)
 		app.grpcServices[svc] = service
 		return service, nil
-	case v2beta1.Node:
+	case grpcserver.NodeV2Beta1:
 		service := v2beta1.NewNodeService(app.host, app.mesh, app.clock, app.syncer)
 		app.grpcServices[svc] = service
 		return service, nil
-	case v2beta1.Layer:
+	case grpcserver.LayerV2Beta1:
 		service := v2beta1.NewLayerService(app.apiDB)
 		app.grpcServices[svc] = service
 		return service, nil
-	case v2beta1.LayerStream:
+	case grpcserver.LayerStreamV2Beta1:
 		service := v2beta1.NewLayerStreamService(app.apiDB)
 		app.grpcServices[svc] = service
 		return service, nil
-	case v2beta1.Transaction:
+	case grpcserver.TransactionV2Beta1:
 		service := v2beta1.NewTransactionService(app.apiDB, app.conState, app.syncer, app.txHandler, app.host)
 		app.grpcServices[svc] = service
 		return service, nil
-	case v2beta1.TransactionStream:
+	case grpcserver.TransactionStreamV2Beta1:
 		service := v2beta1.NewTransactionStreamService()
 		app.grpcServices[svc] = service
 		return service, nil
-	case v2beta1.Account:
+	case grpcserver.AccountV2Beta1:
 		service := v2beta1.NewAccountService(app.apiDB, app.conState)
 		app.grpcServices[svc] = service
 		return service, nil
@@ -1854,7 +1855,7 @@ func (app *App) startAPIServices(ctx context.Context) error {
 		if err != nil {
 			return err
 		}
-		svc.(*grpcserver.SmesherService).SetPostServiceConfig(app.Config.POSTService)
+		svc.(*v1.SmesherService).SetPostServiceConfig(app.Config.POSTService)
 		if app.Config.SMESHING.Start {
 			if app.Config.SMESHING.CoinbaseAccount == "" {
 				return errors.New("smeshing enabled but no coinbase account provided")
